Zentralisierung der Gameloop
