+++ Frontend/src/components/feature/game/src/components/GameField.vue
@@ -1,11 +1,17 @@
 <script setup lang="ts">
-  import { ref,computed } from 'vue';
+  import { ref,computed,onMounted } from 'vue';
   import PacmanObject from './PacmanObject.vue';
   import PacmanPoints from './PacmanPoints.vue';
   import PacmanPowerUp from './PacmanPowerUp.vue';
   import Ghost from './Ghosts.vue'
 
   const score = ref(0);
+
+  const pacmanRef = ref();
+  const ghostRef = ref();
+
+  let lastMoveTime = 0;
+  const moveInterval = 20;
 
   const grid = ref([
   // 1  2  3  4  5  6  7  8  9  10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28
@@ -62,6 +68,17 @@
     ).filter((cell): cell is { row: number, col: number } => cell !== null);
   });
 
+  function gameLoop(timestamp: number) {
+  if (timestamp - lastMoveTime > moveInterval) {
+    pacmanRef.value?.updatePacmanPosition();
+    ghostRef.value?.updateGhostPosition();
+    lastMoveTime = timestamp;
+  }
+    requestAnimationFrame(gameLoop);
+  } onMounted(() => {
+    requestAnimationFrame(gameLoop);
+  });
+
   function updateGrid({ row, col, value }) { 
     grid.value[row][col] = value;
     if(value == 4||value ==6){
@@ -87,8 +104,8 @@
     </div>
     <div class="pacman-container">
       <img class="pacman-maze" src="@/assets/PacManMaze.png">
-      <PacmanObject class="pacman" :grid="grid" @update-grid="updateGrid"></PacmanObject>
-      <Ghost class="ghost" :grid="grid" @update-grid="updateGrid"></Ghost> 
+      <PacmanObject ref="pacmanRef" class="pacman" :grid="grid" @update-grid="updateGrid"></PacmanObject>
+      <Ghost ref="ghostRef" class="ghost" :grid="grid" @update-grid="updateGrid"></Ghost> 
       <div v-for="(cell, index) in showPoints" :key="`point-${index}`" class="showPoints" :style="{ gridRow: cell.row + 1, gridColumn: cell.col + 1 }">
         <PacmanPoints></PacmanPoints>
       </div>
